Refactor author controller and service. Add NotFoundException.

+++ Library.Application/Services/AuthorService.cs
@@ -1,6 +1,8 @@
 using System.Linq.Expressions;
 using AutoMapper;
+using FluentValidation;
 using Library.Application.Contracts;
+using Library.Application.Exceptions;
 using Library.Application.IServices;
 using Library.Domain.Models;
 using Library.Persistence;
@@ -9,7 +11,8 @@ namespace Library.Application.Services;
 
 public class AuthorService(
     IUnitOfWork unitOfWork,
-    IMapper mapper) : IAuthorService
+    IMapper mapper,
+    IValidator<AuthorRequest> validator) : IAuthorService
 {
     public async Task<List<AuthorResponse>> GetAll(Expression<Func<Author, bool>>? filter = null,
         int pageSize = 0, int pageNumber = 0)
@@ -22,12 +25,21 @@ public class AuthorService(
     public async Task<AuthorResponse> GetById(int id)
     {
         var author = await unitOfWork.AuthorsRepository.GetById(id);
+        
+        if (author is null)
+        {
+            throw new ItemNotFoundException();
+        }
         var authorResponse = mapper.Map<AuthorResponse>(author);
         return authorResponse;
     }
 
     public async Task Create(AuthorRequest authorCreateRequest)
     {
+        var validationContext = new ValidationContext<AuthorRequest>(authorCreateRequest);
+        validationContext.RootContextData["IsCreate"] = true;
+        await validator.ValidateAsync(validationContext);
+        
         var authorToCreate = mapper.Map<Author>(authorCreateRequest);
         await unitOfWork.AuthorsRepository.CreateAsync(authorToCreate);
         await unitOfWork.SaveChangesAsync();
@@ -35,6 +47,11 @@ public class AuthorService(
 
     public async Task Update(AuthorRequest authorUpdateRequest)
     {
+        var validationContext = new ValidationContext<AuthorRequest>(authorUpdateRequest);
+        validationContext.RootContextData["IsUpdate"] = true;
+        validationContext.RootContextData["Id"] = authorUpdateRequest.Id;
+        await validator.ValidateAsync(validationContext);
+        
         var authorToUpdate = mapper.Map<Author>(authorUpdateRequest);
         await unitOfWork.AuthorsRepository.UpdateAsync(authorToUpdate);
         await unitOfWork.SaveChangesAsync();
@@ -42,6 +59,11 @@ public class AuthorService(
 
     public async Task Remove(int id)
     {
+        var author = await unitOfWork.AuthorsRepository.GetById(id);
+        if (author is null)
+        {
+            throw new ItemNotFoundException();
+        }
         await unitOfWork.AuthorsRepository.RemoveAsync(id);
     }
 }
